+++ lib/gds_api/account_api.rb
@@ -89,17 +89,6 @@ class GdsApi::AccountApi < GdsApi::Base
     patch_json("#{endpoint}/api/attributes", { attributes: attributes }, auth_headers(govuk_account_session))
   end
 
-  # Look up the names of a user's attributes
-  #
-  # @param [String] attributes Names of the attributes to check
-  # @param [String] govuk_account_session Value of the session header
-  #
-  # @return [Hash] The attribute names (if present), and a new session header
-  def get_attributes_names(attributes:, govuk_account_session:)
-    querystring = nested_query_string({ attributes: attributes }.compact)
-    get_json("#{endpoint}/api/attributes/names?#{querystring}", auth_headers(govuk_account_session))
-  end
-
   # Get the details of an account-linked email subscription.
   #
   # @param [String] name Name of the subscription